+++ stroom-dashboard/stroom-dashboard-impl/src/main/java/stroom/dashboard/impl/logging/SearchEventLog.java
@@ -20,6 +20,8 @@ import stroom.dashboard.shared.DownloadSearchResultsRequest;
 import stroom.docref.DocRef;
 import stroom.query.api.v2.ExpressionOperator;
 import stroom.query.api.v2.Param;
+import stroom.query.api.v2.QueryKey;
+import stroom.query.api.v2.Result;
 import stroom.query.api.v2.SearchRequest;
 import stroom.query.shared.DownloadQueryResultsRequest;
 
@@ -27,12 +29,15 @@ import java.util.List;
 
 public interface SearchEventLog {
 
-    void search(String type,
+    void search(QueryKey queryKey,
+                String queryComponentId,
+                String type,
                 String rawQuery,
                 DocRef dataSourceRef,
                 ExpressionOperator expression,
                 String queryInfo,
                 List<Param> params,
+                List<Result> results,
                 Exception ex);
 
     default void downloadResults(DownloadSearchResultsRequest downloadSearchResultsRequest,