+++ stroom-query/stroom-query-impl/src/main/java/stroom/query/impl/QueryServiceImpl.java
@@ -44,6 +44,7 @@ import stroom.query.api.v2.QueryKey;
 import stroom.query.api.v2.ResultRequest;
 import stroom.query.api.v2.ResultRequest.ResultStyle;
 import stroom.query.api.v2.SearchRequest;
+import stroom.query.api.v2.SearchRequestSource;
 import stroom.query.api.v2.SearchResponse;
 import stroom.query.api.v2.TableResultBuilder;
 import stroom.query.api.v2.TableSettings;
@@ -626,10 +627,12 @@ class QueryServiceImpl implements QueryService, QueryFieldProvider {
 
         QueryKey queryKey = searchRequest.getQueryKey();
         final String query = searchRequest.getQuery();
+        Exception exception = null;
+        SearchRequest mappedRequest = null;
 
         if (query != null) {
             try {
-                final SearchRequest mappedRequest = mapRequest(searchRequest);
+                mappedRequest = mapRequest(searchRequest);
                 LOGGER.debug("searchRequest:\n{}\nmappedRequest:\n{}", searchRequest, mappedRequest);
 
                 // Perform the search or update results.
@@ -643,31 +646,10 @@ class QueryServiceImpl implements QueryService, QueryFieldProvider {
                     // Add this search to the history so the user can get back to this
                     // search again.
                     storeSearchHistory(searchRequest);
-
-                    // Log this search request for the current user.
-                    searchEventLog.search(
-                            "StroomQL Search",
-                            searchRequest.getQuery(),
-                            mappedRequest.getQuery().getDataSource(),
-                            mappedRequest.getQuery().getExpression(),
-                            searchRequest.getQueryContext().getQueryInfo(),
-                            searchRequest.getQueryContext().getParams(),
-                            null);
                 }
-
             } catch (final TokenException e) {
+                exception = e;
                 LOGGER.debug(() -> "Error processing search " + searchRequest, e);
-
-                if (queryKey == null) {
-                    searchEventLog.search(
-                            "StroomQL Search",
-                            searchRequest.getQuery(),
-                            null,
-                            null,
-                            searchRequest.getQueryContext().getQueryInfo(),
-                            searchRequest.getQueryContext().getParams(),
-                            e);
-                }
 
                 result = new DashboardSearchResponse(
                         nodeInfo.getThisNodeName(),
@@ -679,18 +661,8 @@ class QueryServiceImpl implements QueryService, QueryFieldProvider {
                         null);
 
             } catch (final RuntimeException e) {
+                exception = e;
                 LOGGER.debug(() -> "Error processing search " + searchRequest, e);
-
-                if (queryKey == null) {
-                    searchEventLog.search(
-                            "StroomQL Search",
-                            searchRequest.getQuery(),
-                            null,
-                            null,
-                            searchRequest.getQueryContext().getQueryInfo(),
-                            searchRequest.getQueryContext().getParams(),
-                            e);
-                }
 
                 result = new DashboardSearchResponse(
                         nodeInfo.getThisNodeName(),
@@ -700,6 +672,23 @@ class QueryServiceImpl implements QueryService, QueryFieldProvider {
                         null,
                         true,
                         null);
+            } finally {
+                if (queryKey == null) {
+                    searchEventLog.search(
+                            searchRequest.getQueryKey(),
+                            NullSafe.get(
+                                    searchRequest,
+                                    QuerySearchRequest::getSearchRequestSource,
+                                    SearchRequestSource::getComponentId),
+                            "StroomQL Search",
+                            searchRequest.getQuery(),
+                            NullSafe.get(mappedRequest, SearchRequest::getQuery, Query::getDataSource),
+                            NullSafe.get(mappedRequest, SearchRequest::getQuery, Query::getExpression),
+                            searchRequest.getQueryContext().getQueryInfo(),
+                            searchRequest.getQueryContext().getParams(),
+                            NullSafe.get(result, DashboardSearchResponse::getResults),
+                            exception);
+                }
             }
         }
 
@@ -762,7 +751,9 @@ class QueryServiceImpl implements QueryService, QueryFieldProvider {
     }
 
     @Override
-    public ContextualQueryHelp getQueryHelpContext(final String query, final int row, final int col) {
+    public ContextualQueryHelp getQueryHelpContext(final String query,
+                                                   final int row,
+                                                   final int col) {
         return securityContext.useAsReadResult(() -> {
             if (NullSafe.isBlankString(query) || (row == 0 && col == 0)) {
                 return EMPTY_QUERY_CONTEXT;