+++ stroom-state/stroom-planb-impl/src/main/java/stroom/planb/impl/db/AbstractDb.java
@@ -402,7 +402,7 @@ public abstract class AbstractDb<K, V> implements AutoCloseable {
         }
     }
 
-    public void lock(final Runnable runnable) {
+    public void lockCommits(final Runnable runnable) {
         dbCommitLock.lock();
         try {
             runnable.run();