+++ stroom-state/stroom-state-impl/src/test/java/stroom/state/impl/TestTemporalRangedStateDao.java
@@ -27,6 +27,7 @@ import stroom.state.impl.dao.TemporalRangedStateDao;
 import stroom.state.impl.dao.TemporalRangedStateRequest;
 import stroom.state.impl.dao.TemporalState;
 
+import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 
 import java.nio.ByteBuffer;
@@ -38,88 +39,89 @@ import java.util.concurrent.atomic.AtomicInteger;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+@Disabled
 class TestTemporalRangedStateDao {
-
-    @Test
-    void testDao() {
-        ScyllaDbUtil.test((sessionProvider, tableName) -> {
-            final TemporalRangedStateDao stateDao = new TemporalRangedStateDao(sessionProvider, tableName);
-
-            Instant refTime = Instant.parse("2000-01-01T00:00:00.000Z");
-            insertData(stateDao, refTime, "test", 100, 10);
-
-            final TemporalRangedStateRequest stateRequest =
-                    new TemporalRangedStateRequest("TEST_MAP", 11, refTime);
-            final Optional<TemporalState> optional = stateDao.getState(stateRequest);
-            assertThat(optional).isNotEmpty();
-            final TemporalState res = optional.get();
-            assertThat(res.key()).isEqualTo("11");
-            assertThat(res.effectiveTime()).isEqualTo(refTime);
-            assertThat(res.typeId()).isEqualTo(StringValue.TYPE_ID);
-            assertThat(res.getValueAsString()).isEqualTo("test");
-
-            final FieldIndex fieldIndex = new FieldIndex();
-            fieldIndex.create(RangedStateFields.KEY_START);
-            final AtomicInteger count = new AtomicInteger();
-            stateDao.search(new ExpressionCriteria(ExpressionOperator.builder().build()), fieldIndex, null,
-                    v -> count.incrementAndGet());
-            assertThat(count.get()).isEqualTo(100);
-        });
-    }
-
-    @Test
-    void testRemoveOldData() {
-        ScyllaDbUtil.test((sessionProvider, tableName) -> {
-            final TemporalRangedStateDao stateDao = new TemporalRangedStateDao(sessionProvider, tableName);
-
-            Instant refTime = Instant.parse("2000-01-01T00:00:00.000Z");
-            insertData(stateDao, refTime, "test", 100, 10);
-            insertData(stateDao, refTime, "test", 10, -10);
-
-            assertThat(stateDao.count()).isEqualTo(109);
-
-            stateDao.removeOldData(refTime);
-            assertThat(stateDao.count()).isEqualTo(100);
-
-            stateDao.removeOldData(Instant.now());
-            assertThat(stateDao.count()).isEqualTo(0);
-        });
-    }
-
-    @Test
-    void testCondense() {
-        ScyllaDbUtil.test((sessionProvider, tableName) -> {
-            final TemporalRangedStateDao stateDao = new TemporalRangedStateDao(sessionProvider, tableName);
-
-            Instant refTime = Instant.parse("2000-01-01T00:00:00.000Z");
-            insertData(stateDao, refTime, "test", 100, 10);
-            insertData(stateDao, refTime, "test", 10, -10);
-
-            assertThat(stateDao.count()).isEqualTo(109);
-
-            stateDao.condense(refTime);
-            assertThat(stateDao.count()).isEqualTo(100);
-
-            stateDao.condense(Instant.now());
-            assertThat(stateDao.count()).isEqualTo(1);
-        });
-    }
-
-    private void insertData(final TemporalRangedStateDao stateDao,
-                            final Instant refTime,
-                            final String value,
-                            final int rows,
-                            final long deltaSeconds) {
-        final ByteBuffer byteBuffer = ByteBuffer.wrap((value).getBytes(StandardCharsets.UTF_8));
-        for (int i = 0; i < rows; i++) {
-            final Instant effectiveTime = refTime.plusSeconds(i * deltaSeconds);
-            final TemporalRangedState state = new TemporalRangedState(
-                    10,
-                    30,
-                    effectiveTime,
-                    StringValue.TYPE_ID,
-                    byteBuffer);
-            stateDao.insert(Collections.singletonList(state));
-        }
-    }
+//
+//    @Test
+//    void testDao() {
+//        ScyllaDbUtil.test((sessionProvider, tableName) -> {
+//            final TemporalRangedStateDao stateDao = new TemporalRangedStateDao(sessionProvider, tableName);
+//
+//            Instant refTime = Instant.parse("2000-01-01T00:00:00.000Z");
+//            insertData(stateDao, refTime, "test", 100, 10);
+//
+//            final TemporalRangedStateRequest stateRequest =
+//                    new TemporalRangedStateRequest("TEST_MAP", 11, refTime);
+//            final Optional<TemporalState> optional = stateDao.getState(stateRequest);
+//            assertThat(optional).isNotEmpty();
+//            final TemporalState res = optional.get();
+//            assertThat(res.key()).isEqualTo("11");
+//            assertThat(res.effectiveTime()).isEqualTo(refTime);
+//            assertThat(res.typeId()).isEqualTo(StringValue.TYPE_ID);
+//            assertThat(res.getValueAsString()).isEqualTo("test");
+//
+//            final FieldIndex fieldIndex = new FieldIndex();
+//            fieldIndex.create(RangedStateFields.KEY_START);
+//            final AtomicInteger count = new AtomicInteger();
+//            stateDao.search(new ExpressionCriteria(ExpressionOperator.builder().build()), fieldIndex, null,
+//                    v -> count.incrementAndGet());
+//            assertThat(count.get()).isEqualTo(100);
+//        });
+//    }
+//
+//    @Test
+//    void testRemoveOldData() {
+//        ScyllaDbUtil.test((sessionProvider, tableName) -> {
+//            final TemporalRangedStateDao stateDao = new TemporalRangedStateDao(sessionProvider, tableName);
+//
+//            Instant refTime = Instant.parse("2000-01-01T00:00:00.000Z");
+//            insertData(stateDao, refTime, "test", 100, 10);
+//            insertData(stateDao, refTime, "test", 10, -10);
+//
+//            assertThat(stateDao.count()).isEqualTo(109);
+//
+//            stateDao.removeOldData(refTime);
+//            assertThat(stateDao.count()).isEqualTo(100);
+//
+//            stateDao.removeOldData(Instant.now());
+//            assertThat(stateDao.count()).isEqualTo(0);
+//        });
+//    }
+//
+//    @Test
+//    void testCondense() {
+//        ScyllaDbUtil.test((sessionProvider, tableName) -> {
+//            final TemporalRangedStateDao stateDao = new TemporalRangedStateDao(sessionProvider, tableName);
+//
+//            Instant refTime = Instant.parse("2000-01-01T00:00:00.000Z");
+//            insertData(stateDao, refTime, "test", 100, 10);
+//            insertData(stateDao, refTime, "test", 10, -10);
+//
+//            assertThat(stateDao.count()).isEqualTo(109);
+//
+//            stateDao.condense(refTime);
+//            assertThat(stateDao.count()).isEqualTo(100);
+//
+//            stateDao.condense(Instant.now());
+//            assertThat(stateDao.count()).isEqualTo(1);
+//        });
+//    }
+//
+//    private void insertData(final TemporalRangedStateDao stateDao,
+//                            final Instant refTime,
+//                            final String value,
+//                            final int rows,
+//                            final long deltaSeconds) {
+//        final ByteBuffer byteBuffer = ByteBuffer.wrap((value).getBytes(StandardCharsets.UTF_8));
+//        for (int i = 0; i < rows; i++) {
+//            final Instant effectiveTime = refTime.plusSeconds(i * deltaSeconds);
+//            final TemporalRangedState state = new TemporalRangedState(
+//                    10,
+//                    30,
+//                    effectiveTime,
+//                    StringValue.TYPE_ID,
+//                    byteBuffer);
+//            stateDao.insert(Collections.singletonList(state));
+//        }
+//    }
 }